Common implementation for REST get_agreement
