+++ core/market/decentralized/src/rest_api/provider.rs
@@ -5,6 +5,7 @@ use std::sync::Arc;
 use ya_client::model::market::{Offer, Proposal};
 use ya_service_api_web::middleware::Identity;
 
+use super::common::*;
 use crate::market::MarketService;
 
 use super::{
@@ -148,17 +149,4 @@ async fn terminate_agreement(
     _id: Identity,
 ) -> HttpResponse {
     HttpResponse::NotImplemented().finish()
-}
-
-#[actix_web::get("/agreements/{agreement_id}")]
-async fn get_agreement(
-    market: Data<Arc<MarketService>>,
-    body: Path<PathAgreement>,
-    id: Identity,
-) -> impl Responder {
-    let agreement_id = body.into_inner().agreement_id;
-    market
-        .get_agreement(&agreement_id, &id)
-        .await
-        .map(|agreement| HttpResponse::Ok().json(agreement))
 }
