+++ core/market/decentralized/src/rest_api/requestor.rs
@@ -8,6 +8,7 @@ use ya_service_api_web::middleware::Identity;
 
 use crate::market::MarketService;
 
+use super::common::*;
 use super::{
     PathAgreement, PathSubscription, PathSubscriptionProposal, ProposalId, QueryTimeout,
     QueryTimeoutMaxEvents,
@@ -131,19 +132,6 @@ async fn create_agreement(
         .create_agreement(id, &proposal_id, valid_to)
         .await
         .map(|agreement_id| HttpResponse::Ok().json(agreement_id))
-}
-
-#[actix_web::get("/agreements/{agreement_id}")]
-async fn get_agreement(
-    market: Data<Arc<MarketService>>,
-    body: Path<PathAgreement>,
-    id: Identity,
-) -> impl Responder {
-    let agreement_id = body.into_inner().agreement_id;
-    market
-        .get_agreement(&agreement_id, &id)
-        .await
-        .map(|agreement| HttpResponse::Ok().json(agreement))
 }
 
 #[actix_web::post("/agreements/{agreement_id}/confirm")]
